Clean up and move ar guest test
